Add files via upload
You can modify the usernames list with the list of usernames you want to check, and change the url_template variable to the format of the URLs you want to check against.

Note that this is just a simple example and there are many more advanced username enumeration tools and libraries available in Python. Also, be aware that username enumeration can be used as part of social engineering attacks and should be used responsibly and with permission.